+++ core/src/main/kotlin/org/evomaster/core/sql/SqlInsertBuilder.kt
@@ -18,7 +18,8 @@ import org.evomaster.core.logging.LoggingUtil
 
 
 class SqlInsertBuilder(
-    schemaDto: DbInfoDto,
+    public val schemaDto: DbInfoDto,
+
     private val dbExecutor: DatabaseExecutor? = null
 ) {