+++ e2e-tests/spring-rest-bb/src/main/kotlin/com/foo/rest/examples/bb/exampleobject/ExampleObjectDto.kt
@@ -1,0 +1,14 @@
+package com.foo.rest.examples.bb.exampleobject
+
+class ExampleObjectDto {
+
+    var id: String? = null
+
+    var x: Int? = null
+
+    var b: Boolean? = null
+
+    var y: Double? = null
+
+    var other: OtherDto? = null
+}