+++ e2e-tests/spring-rest-bb/src/main/kotlin/com/foo/rest/examples/bb/exampleobject/OtherDto.kt
@@ -1,0 +1,8 @@
+package com.foo.rest.examples.bb.exampleobject
+
+class OtherDto {
+
+    val name: String? = null
+
+    val x: Int? = null
+}