+++ release_notes.md
@@ -2,6 +2,14 @@
 
 Under development in `master` branch.
 
+### New Features
+- Support for object in example/examples in OpenAPI schemas.
+
+### Addressed GitHub Issues
+- #1171: IllegalStateException: only support Map with String key in EvoMaster
+- #1159: EvoMaster crashed when reading em.yaml file 
+
+
 # Version: 3.4.0
 
 ### New Features